Move formatter function to their own file
